Extract formatBytes function to separate utility module
